x86/hyperlaunch: efi detect and load hyperlaunch config
Detect if a FDT is present and load it as the hyperlaunch configuration.
This commit is focused only on enabling dom0less configuration booting,
so may not work with general hyperlaunch configuration.

This moves the arch device tree detection x86 stub that will be used by
the domain builder.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>

fixup: efi_arch_handle_cmdline - add NULL image name - it is unused.
Fix xen cmdline with valid hyperlaunch config.
Centralized with a single check for not finding a DT path.
Use IS_ENABLED(CONFIG_DOM0LESS_BOOT)

Signed-off-by: Jason Andryuk <jason.andryuk@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>