x86/hyperlaunch: Check DT for Hyperlaunch in EFI entry
Use has_hyperlaunch_node() in efi_arch_use_config_file() to ensure a
valid Hyperlaunch DT is present before setting Hyperlaunch mode.

coreboot + u-boot seems to always pass a DT through the config table,
even it not specifying one for Hyperlaunch.  If only the presence of a
DT is used, we don't follow the xen.cfg path.  Later, boot fails when
Xen determines that the DT is missing the Hyperlaunch config.

So the presence of a DT is not enough - it has to be a Hyperlaunch DT.
Otherwise we can't use the xen.cfg path.

Signed-off-by: Jason Andryuk <jason.andryuk@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>