x86/hyperlaunch: Seed the grant table of all domains.
Grant the xenstore domain access to all xenstore pages of all domains.

Without this, xenstore cannot access the xenstore pages.

While at it, avoid seeking the xenstore domain when allocating the
evtchn, as its domid is now already present in the boot_domain.

The new if ( IS_ENABLED() ) check exists to let DCE remove the functions
on non-dom0less builds and avoid a missing definition.

Fixes: c479b1eb3e54("x86/hyperlaunch: allocate xenstore for domu"
Signed-off-by: Alejandro Vallejo <alejandro.garciavallejo@amd.com>
Reviewed-by: Jason Andryuk <jason.andryuk@amd.com>