x86/hyperlaunch: allocate console for domu
During domU construction, a page of memory and an event channel must
be setup for the console connection. In this commit, a page from the
special page region of domU is setup as the console page along with an
event channel. The page address and event channel are published in the
HVM parameters, so they may be published in Xenstore once it is online.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>
Signed-off-by: Alejandro Vallejo <alejandro.garciavallejo@amd.com>
Reviewed-by: Jason Andryuk <jason.andryuk@amd.com>