tools/init-dom0less: Add console initialisation in xenstore
Gate it on the evtchn port being non-zero, because only x86 has Xen
prepopulating the HVM params and seeding the grant tables for the time
being.

Assume the console is in dom0/hwdom, for lack of better readily
available information. We could make it configurable from the command
line, but chances are we'll always start it with zero.

Ensure that when seeding the grant table for a domain with late-alloc
xenstore page we seed the grant table with the correct gfn for the
console in order to avoid overriding a potentially active grant.

Xen doesn't currently prepare the console enhancements for ARM, but
better safe than sorry in case we ever port them.

Signed-off-by: Alejandro Vallejo <alejandro.garciavallejo@amd.com>
Reviewed-by: Jason Andryuk <jason.andryuk@amd.com>