xen/arm: introduce mali_aw
This patch introduces the mali_aw field to the Xen domain
configuration and internal domain structure. The Access Window (AW)
ID is the only required configuration for assigning GPU resources to
a VM in the Mali-G78AE virtualization framework.

The mali_aw field tracks which AW is assigned to a given domain,
enabling the arbiter and resource group to correctly route GPU
requests.

Signed-off-by: Luca Miccio <luca.miccio@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>