xen/arm: Temporary solution to treat CMA as static shared memory region
CONFIG_STATIC_SHM needs to be enabled, so add it to defconfig.

Problem:
Xen does not handle /reserved-memory nodes, especially when the node has
reusable property which means Linux can use it for its own RAM (and
shared structures with Xen) if device does not claim access to it.

Worflow:
When Xen detects a /reserved-memory child node with compatible
"shared-dma-pool", it uses static shared memory code to set up a region
for dom0 as owner (region is mapped in P2M as normal RAM).

Signed-off-by: Michal Orzel <michal.orzel@amd.com>
Reviewed-by: Michal Orzel <michal.orzel@amd.com>