x86/hvm: make standard vga emulation optional
The virtual Standard VGA is not enabled for PVH Guests and not in safety
scope, where only PHV Guests are allowed, so having virtual Standard VGA
enabled unconditionally leaves unjustified gaps in terms of coverage (MISRA
Rule 2.1).

This patch introduces CONFIG_VVGA to allow the Standard VGA emulation to be
compiled out.

Signed-off-by: Grygorii Strashko <grygorii_strashko@epam.com>
Reviewed-by: Jason Andryuk <jason.andryuk@amd.com>