common: enable sending VIRQ_CONSOLE to non hardware domains
The linux "hvc" driver binds to the VIRQ_CONSOLE as a part of
initializing the xen console for input.

To enable linux with console I/O for x86_64 predefined domain tests,
make VIRQ_CONSOLE a per domain VIRQ instead of global, and for x86_64
also send it to non hardware domains to avoid linux boot failure.

Signed-off-by: Victor Lira <victorm.lira@amd.com>
Reviewed-by: Jason Andryuk <jason.andryuk@amd.com>
Reviewed-by: Grygorii Strashko <grygorii_strashko@epam.com>