hvm/mtrr: introduce reset function for cacheattr
Reinitialize MTRR.

TODO: If MTRR default type is not enabled, defaults to UC. Hence,
hyperlaunched guests whose MTRR def type is not updated by toolstack
or hvmloader to WB, do not conform to the behavior stated in
docs/misc/pvh.pandoc Hence, that guarantees that RAM regions will
always have the WB cache type. This needs to be fixed in another patch.

Signed-off-by: Xenia Ragiadakou <xenia.ragiadakou@amd.com>
Reviewed-by: Alejandro Vallejo <alejandro.garciavallejo@amd.com>