ioreq: introduce ioreq_server_{enable/disable}_all()
The function ioreq_server_disable_all() disables all the ioreq
servers of a domain, clears sv->pending and ioreq and buffered
ioreq pages. In the context of domain reset, disabling the ioreq
servers prevents invalidation ioreqs, triggered by p2m changes,
to block the vcpu in case of self-reset. For x86, ioreq pages are
reset by arch_ioreq_server_disable(s) but this is not the case for
arm, so they are cleared explicitely.

The function ioreq_server_enable_all() reinits the ioreq pages
and enables all the ioreq servers of a domain.

Signed-off-by: Xenia Ragiadakou <xenia.ragiadakou@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>