xen/arm: Check for all CMA regions in temporary fix
The initial workaround for CMA targeted shared-dma-pool instead of the
standard linux,cma-default. This was corrected in commit 89aebf, which
switched to using linux,cma-default.

However, multiple CMA regions may be defined in the device tree, either
via linux,cma-default or as shared-dma-pool nodes marked reusable.

This patch ensures that all such CMA regions are properly handled by
creating a corresponding shared memory bank for each one, regardless of
how it is defined.

Fixes: 47be7d3b2ff5 ("xen/arm: Temporary solution to treat CMA as static shared memory region")
Signed-off-by: Stefano Stabellini <stefano.stabellini@amd.com>
Signed-off-by: Luca Miccio <luca.miccio@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>
Reviewed-by: Stewart Hildebrand <Stewart.Hildebrand@amd.com>