x86: Migrate switch vendor checks to cpu_vendor()
In the single-vendor case (with no unknown vendor) all switches turn
to a branchless straight line through the single compiled-in case.

In the multivendor case where some vendors are disabled the VRP pass
ensures those branches are eliminated.

It's hard to tell the VRP pass zero is also unreachable when
CONFIG_UNKNOWN_CPU_VENDORS is disabled, so default cases tend to remain.

Signed-off-by: Alejandro Vallejo <alejandro.garciavallejo@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>