x86: Have x86_emulate/ implement the single-vendor optimisation
Open code the vendor check through the policy as a one-off. The emulator
embeds amd_like() in macros and is called in MANY places. Using a
local variable (cp->x86_vendor) makes it a lot smaller (300-400 bytes
smaller). So treat this as the exception it is and let it use the policy
rather than boot_cpu_data.

This deviation is STRICTLY for codegen improvements when not using a
single vendor. It compiles to the same constant under the single vendor
optimisation.

Also keep the current behaviour of using the policy vendor when
compiled for userspace, where cross-vendor configurations are expected.

Not a functional change.

Signed-off-by: Alejandro Vallejo <alejandro.garciavallejo@amd.com>
Reviewed-by: Victor Lira <victorm.lira@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>