xen/{arm,x86}: Disable CONFIG_DEBUG
We want safety defconfigs to be as close to the final configuration as
possible. For that, we want to use Xen release build. In order for the
test that rely on Xen debug logs to pass, use the same logging thresholds
by using the DEBUG macros also when CONFIG_VERBOSE_DEBUG is set. Otherwise we
would need to modify the specs and tests to append a long Xen cmdline
to set thresholds manually. With CONFIG_DEBUG=n, the null scheduler is no
longer enabled by default, so enable it explicitly.

Signed-off-by: Michal Orzel <michal.orzel@amd.com>
Reviewed-by: Ayan Kumar Halder <ayan.kumar.halder@amd.com>