xen: Introduce GICv3 SGI tests
Introduce the tests to confirm that:
1. A cpu can send SGI 0 to itself
2. A cpu can send SGI 0 to another specific CPU
3. A cpu can send SGI 0 to all the other CPUs
4. A cpu can send SGI 1 to another CPU

These tests aim to test Xen has configured the GIC correctly to use SGIs.
Thus, the tests invoke specific APIs of GIC driver.

Also, introduce a config CONFIG_AMD_TESTS which enables these tests. This
tests should be disabled for production builds. They are enabled for
running the validation pipeline and measuring the coverage.
These tests are executed when Xen boots. They are aimed at validing the
internal interfaces such as Xen's interface with hardware, firmware,
bootloader, etc.

Also, introduce an integer command line parameter "gic-test". By default, it
is set to 0 which means no tests are enabled.
For running SGI tests, "gic-test" should be set to 1. In future if we add
tests for distributer, ITS, LPI, etc, then we can use different numbers.
Thus, each number denotes a functionality of GICv3 which can be tested
independently and within a single boot of Xen.

In this way, we ensure that the tests to validate SGIs do not impact any other
tests.

In order to keep all the AMD_TESTS together in the binary, we have introduced
a separate section "initcallboottest". All the tests are registered using
__initcallboottest. During the bootup of each core, Xen invokes
do_init_boottests() to run the these tests. All these tests are invoked
before Xen creates the domains (in case of primary core) or runs the idle
loop (in case of secondary
core).

Signed-off-by: Ayan Kumar Halder <ayan.kumar.halder@amd.com>
Signed-off-by: Michal Orzel <michal.orzel@amd.com>
Reviewed-by: Grygorii Strashko <grygorii_strashko@epam.com>